+++ frontend/app/app.component.ts
@@ -1,7 +1,15 @@
 import { Component } from '@angular/core';
-
 @Component({
   selector: 'my-app',
-  template: `<h1>Hello {{name}}</h1>`,
+  template: `
+    <h1>{{title}}</h1>
+   <nav>
+     <a routerLink="/dashboard">Dashboard</a>
+     <a routerLink="/books">Books</a>
+   </nav>
+    <router-outlet></router-outlet>
+  `
 })
-export class AppComponent  { name = 'Angular'; }
+export class AppComponent {
+  title = 'Project Phoera';
+}