+++ frontend/app/mock-books.ts
@@ -1,0 +1,6 @@
+import {Book} from "./book";
+
+export const BOOKS: Book[] = [
+  { id: 1, title: 'Flight from the Dark' },
+  { id: 2, title: 'The Jungle of Horrors' }
+];